sentient-brain-surreal/requirements.txt
# Core Framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
aiohttp>=3.9.5
python-dotenv>=1.0.1

# Multi-Agent Framework & LLM Integration
pydantic>=2.5.0
langgraph>=0.2.0
groq>=0.4.0

# Embeddings & AI Models
google-genai>=1.20.0
sentence-transformers>=3.0.0
transformers>=4.40.0
torch>=2.0.0

# Code Parsing & AST
tree-sitter>=0.21.0
tree-sitter-python>=0.21.0
tree-sitter-javascript>=0.21.0
tree-sitter-typescript>=0.21.0

# File Processing & Monitoring
watchdog>=4.0.0
PyYAML>=6.0.1

# Document Processing
markdown>=3.6.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# Testing
pytest>=8.2.2
pytest-asyncio>=0.23.7
